requirements.txt
customtkinter>=5.2.0
pytesseract>=0.3.10
Pillow>=9.0.0
